fixes https://github.com/tmssoftware/tms-smartsetup/issues/219
fixes for compiling dpr in other delphis that are not D7 (which was what Deget originally supported). The fix has been staying in my disk forever, because I am not 100% sure it won't break some other tool like tmsdev, but I think it is time to try it :)
Still it should affect versions like D2006 which I am not sure anyone is using.