Adds the ability to use packages that support more than one IDE. Fixes https://github.com/tmssoftware/tms-smartsetup/issues/246 using option 4.
If a package supports for example D11+ (by using libsuffix auto), the dcus can't go at the same folder for D11 and D12, as it would crash. This fix add an extra folder with the IDE Name, when a project has packages that support more than 1 IDE.
For example, the dcus could go to:
\D11+\d11\Win32\Debug\
and
\D11+\d12\Win32\Debug\

This way dcus for d11 and d12 will not go to the same place.