Removed BaseDir from DelphiPackageInfo
It wasn't being used for anything, and it made code more confusing since you had to specify a parameter that has no meaning.
When adapting this to support exes, BaseDir usage was replaced by OrigPackageDirectory. From what I see in the history, this was because when using BaseDir we concatenated with a hardcoded "packages" folder which didn't apply for exes that aren't in packages folder.
In any case, it wasn't used anymore, so it was just making stuff more complex without any upside