Adds an option to ignore TargetPlatforms in the dproj. Fixes https://github.com/tmssoftware/tms-smartsetup/issues/247
Also restructures a little the sections in the YAML. Since it is possible that we will keep adding stuff to configure packages (like the files that go inside every package), I think it is better to move those options to a group: "Package options".
This doesn't break existing yamls, only changes stuff that we had changed after latest official release v1.10.