Added commands: ServerAdd, ServerEnable, ServerList and ServerRemove, as discussed in https://github.com/tmssoftware/tms-smartsetup/issues/245
Note: I didn't add a serverdisable command, to avoid one extra command. You can disable a server with "tms enable community false" or just editing the yaml. Of course, if needed, we can add a server-disable command.

Also as discussed, we removed the hardcoded urls from the yaml for default servers. So we can change them later and it is not needed to update yamls all over the world.

the third-party server was renamed to community server
tmsServer protocol is now just "api"