Fix to compile JCL
DPKs in JEDI JCL all have a hanging ";" at the end, like:
requires
  rtl
  ;

contains
My guess is that they generate them with some tool. But well, this is a valid dpk, but we weren't able to parse it.
After this changes JCL is compiling fine. Of course, some other badly formatted dpks could still fail, but a "tms install *" is working fine including all tms products and all open source in the registry.