removes the hardcoded urls for the api protocol.
This is working fine in most cases (list-remote will list all api sources, tms credentials will ask the credentials once for each server, etc), but to avoid breaking changes, the fetch part only accepts one api server. (multiple github servers, and even multiple local servers are fine). In particular, the method FindApiServer checks that there is only one api server enabled.

Some day we will need to re-architect the full fetch process so we unify github and api requests and hopefully simplify the full code (which will never be simple because of the dependencies). But this will imply a lot of code changes and potential new bugs, so for now we will leave it with the restriction of not more than one api server.