Improved logic for enabling servers. Before when we did a server-enable community and the list of servers was empty, it would disable tms
This is because when the list is empty, we have 3 predefined servers. But when we
add a server, it is no longer empty, so the other 2 predefines servers aren't there and so are disabled.
Now when we add a server and the list is empty, we add all the predefined servers anyway.