Added option to use a zip file as "repository" besides a GIT or SVN repo
The current repo options that we had (SVN and GIT) will always get the trunk version of the component (At least until we implement versioning). But what if instead of trunk, I wanted say release 1.0.0 of aws SDK?  Just that version, not a newer one.

Now with this commit you can do:
tms repo-register zipfile https://github.com/landgraf-dev/aws-sdk-delphi/archive/refs/tags/1.0.0.zip
tms install *aws*

And it will install with this particular release: https://github.com/landgraf-dev/aws-sdk-delphi/releases/tag/1.0.0
(and yes, it is already working :))

But this new "zipfile" protocol might have other uses too. Basically, it allows you to get any smartsetup bundle that is hosted in a website. It will just download the zip from that website, then unzip it and build it as a normal product.

You might create a registry with zip files for your organization with specific bundles for each product, and get them from there.

The only restriction is that the webserver has to return [etags](https://developer.mozilla.org/en-US/docs/Web/HTTP/Reference/Headers/ETag) so we can know if there was an update to the component or not (and not just download blindly every time you want to update). GitHub at least returns etags, so any zip/.tar.ztd hosted in github should work.