Renamed the "github" protocol for open-source servers to zipfile
It brings consistency with the "ZipFile" protocol when registering a repo, and, there is nothing github-specific about the protocol. Any website that can serve a zip file, and supports ETAGS, can be a server.
We also plan to add support for "shared folders" to this protocol. So you could use a url like file:\\mysharedfolder\repos.zip
GitHub was the wrong name.