Added a "company name" files to the product definition.
The value is necessary for when we automatically creating packages, to include the version info in the bpls created.
If empty, well, the packages will have empty version info, but here we give you the choice to provide relevant info so when you right-click the bpl and look at properties you can see them.

It might also be displayed in tmsgui, not sure if that matters.

Company name isn't unlikely to change, so no big deal with it getting out of synch. And if it changes and they are using tms build, well, they better update this file.