Changed the section on "registry keys" so "value" is an array instead of single keys
What we had before was "invalid yaml", even if it was ok for our parser. You can't have a repeated key in a section, and VSCode would complain and red-underline it.
So now it works as expected in yaml. To be backward compatible and still be able to read old webcore files, we still allow the "value:" syntax instead of "-value". A new property ArraysCanBeKeys allows them both (this property is not to be used in the future, only for backwards compat)