Added "registry" and "filelinks" to the skip registering options.
Before, they were all under "packages", but they are different. Specially, before we disabled all registering for exes, but now we only disable the package registering part. (A exe can register itself in the start menu, for example).

But if filelinks and registry were part of "SkipRegistering.Packages", then you couldn't register them in an exe.

Note: This might break existing code: If you had a "skip register: [packages]", not it should be "skip register: [packages, registry, filelinks]". To avoid similar breaks in the future when we add more values to the enum, we now also enable the inverse: "skip register: [all, -help] will skip registering all except the help. If we add a new value to the enum, it will still keep skipping everything except help