Megafolders should be working, closes https://github.com/tmssoftware/tms-smartsetup/discussions/193
This is the second part of the implementation:
  * First part dealt with synching the dcus in the megafolders with the actual ones.
  * This part deals with adding the megafolders to the library path, instead of the folders for every product.

This second part ended up more complex than what I thought, and in fact more complex than the fist part. Because registering all megafolders at the start, kind of like we change the Windows Path, which was the original idea, turned out to be complex, since you can be writing to an alternative registry key, you don't know which delphi versions you will have to register the paths either (or the platforms).

So now we register them when we register the product itself. This makes registering simple: instead of registering `..pathtoproduct...` we register `megafolder` instead.
But there are 2 drawbacks:
  1. The product registration is done in threads, so many products might want to register the same megafolder at the same time.
  2. Uninstalling becomes harder, as we now need to know if any other product is also using the same megafolder before we can remove it.

For 1. I used a TMonitor(MegafoldersLock), so no thread clash
For 2. there is a separate uninstall folder for megafolders. Where you uninstall a product, you can check if other products have the same folders, and if yes, do not unregister. Only the last product unregistering will unregister the megafolder.

Maybe there is a cleaner approach, but this one seems like working, and fixes the "too long library path" for users that have the problem. For the other users (hopefully most), it all remains unchanged.