Bug Fix: tms credentials failed in an empty folder, telling you to run either tms config or tms credentials
When we added the support for multiple servers (even if currently we support only one api server), we changed the code of credentials to loop over all existing api servers and ask the credentials for each. But this required access to Config, which would complain about not being a valid folder. Now if we are not in a valid folder, we just ask credentials for the builtin "tms" server.