If a zip file has an empty top folder, like all files are inside  "aws-sdk-delphi-1.0.0" folder, we remove it.
The extra folder only adds to the library path, and it is not needed. But the zip files we get from GitHub have this top folder added, so if possible, we just get rid of it. Note that if we don't this will still work, just with another extra folder.