Updates ZSTD to version 1.5.7
Also changed the way to create the .o files:

Before we used C++ builder (modern, which is clang), but that lead to very slow compression speeds for some reason. And so, we had to use the dll to compress. Now, .o are compiled with mingw, and compression speed is similar to the dll, so it makes no more sense to use libzstd.dll in smartbuild or tmsdev for compression.

Note: this still depends on which makes this fail in windows server 2012 (...). We could try to use System.Win.Crtl instead, and it seems to work, but crashes with big data. Could be looked at.