Changes tms info so it will work in an empty folder. Same for credentials, which already did, but now it is simpler code.
Credentials now will also assume "tms" as server if none is specified, and go back to returning only "{code/email}" instead of
"{server:{code/email}}" as to not break compat with v1. As "tms" is the default there is no way to not specify a server anymore, so you don't need the server back in the answer.