Fixes https://github.com/tmssoftware/tms-smartsetup/issues/265
The system was written without considering a "-server" parameter. So it always loaded all products from all VCS servers, and to add support for -server we filtered the ones that had the server in -server.
But it was filtering only in json, the line was missing from text.

We could have fixed it by adding the filtering in text output too, but the full thing needed a rework to support -server. Because as it was, it always loaded all products from the VCS servers, even if we only wanted products form tms.

So now, we do the filtering at the right place, and if we provide -server, we don't load products from any other product than that server. I removed the filtering at the json code, as the products now are already filtered. And text will work automatically now too.