Adds the ability to read a "base" tms.config.yaml from the location in the environment variable TMS_STARTING_CONFIGURATION
This allows us to define a "common starting point" for all tests. Say, we can define in the file pointed by TMS_STARTING_CONFIGURATION that the Delphi version used for the tests will be delphi 12. Then the individual tests don't specify a delphi version, but still only run in delphi 12. When we move this to support Delphi13, we only need to cahnge the "base" tms.config.yaml, and all tests will change, even if they have their own. (as long as they don't specify the delphi version in their yaml). Current flow is as follow:

* We start with an empty config with default values.
* If there is a TMS_STARTING_CONFIGURATION env var defined, we load the config from there, overriding the defaults
* If there is a tms.config.yaml, we load the config from there, overriding what was in the configuration.
* If we specify a `-p:...` parameter in the command line, we load that too overwriting what was before. This has the maximum priority.

This commit also has changes to allow you to define a server twice, as long as it is when loading different files. So you could define a server in the base config, and then redefine it again in the real config. But you still can't define it twice in any config file, to avoid errors.