Added test to reproduce https://support.tmssoftware.com/t/c-builder-access-violation-with-static-linking/25711
Sadly the test passes, because the bug only happens when compiling from the IDE (!)
I'll create an issue about it when I find out more. But the problem seems to be the IDE passes the path of the localized stuff to the linker, while msbuild doesn't pass it.

In any case, we pass our own paths, and we currently ignore translated library paths anyway.