Fixes compilation of map2pdb
The "extra defines" section which we used in map2pdb to remove the "madExcept" define, broke when we removed the `/p:Dcc_Define`" in the command line (see https://github.com/tmssoftware/smartsetup/commit/c7ddcea0d6606c1f9fb4e82ddacc7a9936fbda3a). It looks that even when the behavior of /p:Dcc_Define parameter deleting all previous defines gave us a lot of problems, in the implementation of "extra defines' and the "remove" parameter, we were relying in this very characteristic. Now we do it "right" and remove those defines from the dproj

See