Converting the config writer to use the schema.
A new class, BBYamlWriter, should be used to save the yaml now.
Fixed typos in the schema, and also modified Skip-registering so it saves back what it had read. As it was before, if you had `skip register: [All,-StartMenu], it would be saved back as [packages, ...]. While technically the same, the reversed condition was lost.