Added the ability to "nullify" properties from the command line.
This is needed when you have "configuration for product" sections. Let's say I want to remove the "compilation options:debug dcus" entry, so the value is inherited. Currently, you can only set the value to true or false, but not erase it.
Now you can do `tms config-write -p"configuration for product:compilantion options:debug dcus=" and it will remove it.