Added an option to not run pre/post build events, disabling build events by default
I wasn't sure before if disabling build events would really help with security, because if you install a library an then launch Delphi hope is already lost. But still, with build events it might happen that *just compiling* a library could result in a malware attack. This means that just running the tests (that only install/uninstall components) could be a security risk.

With this change, tests should be safe to run. No one of all the libraries we already support have build events anyway. And if it is needed, you can enable them in tms.config.yaml for some select products