Fixed compilation for C++ executables
There was a bug in D12 that we workarounded by passing the include path as "SystemIncludePath" instead of "INcludePath".
But passing the SystemIncludePath as a /p parameter means that the current SystemIncludePath is removed, so we need to pass all standard include paths. Those include a path chat changes in every version ($(BDS)\lib\clang\15.0.7\include;  doe D12, $(BDS)\lib\clang\20\include; for D13), so this would need constant updating.

As the bug that caused this doesn't exist anymore in D13, we go back to just passing "IncludePath" and leaving "SystemIncludePath" alone for D13+