Fixes 'failed to find / extract cached MinGW includes: PathAlreadyExists' error
Now we will run resinator.exe for all delphi versions >=13, so if the windows headers aren't extracted, they will be.
If the headers are already extracted, this is instant so this fix doesn't make tms slower. If they aren't, it will take a second, but that second would happen anyway later when trying to compile.

Running resinator before doing the parallel compilation prevents multiple instance of resinator to try to extrat the same headers at the same time and so crash.
Fixes https://github.com/tmssoftware/smartsetup-registry/issues/5 and https://github.com/tmssoftware/tms-smartsetup/issues/272