Changed TestInstallEverything so it installs into an alternate registry key and uses bds.exe to compile.
The previous test was installing in the "real" delphi registry, which isn't nice when you need ot start delphi to investigate something, as everything will be installed. It is also harder to clean

Now we install to tmstest\everything, which keeps the env cleaner (just remove that key, even if it failed, and all goes away)
Also we were testing that we could compile an app that used some of the installed components after installing, but calling msbuild. This didn't work, as we never launched bds and so EnvOptions.dproj wasn't updated. Now we compile directly with bds instead of msbuild, which also mostly resembles what the user will do.