Refactoring: Split Deget.Compilation in multiple units, (Compilation.Delphi, .Lazarus, Precompiled)
Lazarus currently doesn't do anything since we need to move the line calling lazbuild from Lazarus.Installer to here.
The idea of the refactoring is to make the compilers more cross-plat, so for example a delphi compiler can be Delphi only, but the PrecompiledCompiler is crossplat.

This refactoring is also paving the way to a webcore compiler, which I am not sure if it is really needed or not, so by the moment, I am not commiting the changes to add WebCore