Fixes the case where you want to install `tms install tms.biz.echo:1.0 tms.biz.echo:2.0`
Now the remaining case is dependencies, where kind of the reverse is true. If I want to add tms.biz.aurelius:1.0 tms.biz.bcl:1.0 this musn't fail, even when a dep to tms.biz.bcl:latest is added.

The test for checking 2 deps can't be added in the command line is now passing, added a new test case for the deps.