Fixes the case where you want to install say product A which has Dep B (which would have to be installed at the latest), but you also manually specify a version of B
This case is not trivial, there are many possibilities. At the end, I added a flag to FetchItem (CanBeAnyVersion) that when a dependency sets the version, is set to true. This way we can know when we find conflicting versions if there isa real conflict or the value of the dep can be overwritten.