Added some repositories with dependencies for testing
In our community server all the products we have don't have any dependencies, so we can't test with them.
To avoid creating 5 (or more) dummy github repos for products A, B, C... which depend on each other, I created the packages, and there is a script that automatically creates the repos in the disk with 2 versions (v1.0 and 1.1), and a zipfile server to use them. We can use the repos here to test git products with dependencies.

And btw, while writing the test, I noticed that we don't handle deps at all when downloading from VCS. I could have swear we did it and we had a problem where you couldn't mix deps in api and vcs servers, but I can't find any code. As no VCS product has dependencies, we never noticed it. oh well.