Added dependency handling to Fetching VCS
For some reason it wasn't implemented. Probably because we were waiting for a more general fetching solution that would mix VCL and API, allowing for a API product to have a dep in a VCS product and vice versa.
This implementation allows only for dependencies between products of the same type (and server?)