Fixes the download recycling for handling versions.
In the old system, we had:
  * Last version always goes into Downloads\CurrentVersions
  * Old versions are moved to Downloads\OldVersions and deleted if they pass the maximum allowed.

Now, with versions logic gets mor complex.
  * If I download v1.0, that should be the version in CurrentVersions, even if newer versions exist
  * When we delete, we delete the oldest versions as before, but not v1.0 as it is current.
  * When searching for a file to download, we need to check if it exists in both CurrentVersions and OldVersions. Before, there was no need to check in OldVersions as we could only update, not go back to older versions.

Note: In https://github.com/tmssoftware/tms-smartsetup/issues/285 we discussed the possiblity of just removing the OldDownloads/CurrentDownloads folder since they don't have a lot of use now that "current' might not mean "latest".
But code changes were similar (maybe even less effort to keep the current system), and with this implementation this is non-breaking. Some user could already be backing up the folder "currentDownloads", and if we remove it, the scripts would break. So we just kept the structure as it was.