Fixes a bug where if you fetched a product from the community server and it didn't had a yaml, it wouldn't update after the first fetch
As it was before, when you fetched a product that didn't provide its own tmsbuild.yaml, it would get it from the server (which is just a collection of tmsbuild.yaml).

But then, if you updated the tmsbuild.yaml in the server, it wouldn't be updated, as we give more priority to existing tmsbuild.yaml and there was already one in the disk (the old one from the first fetch). The only way to get the updated yaml was to uninstall the product and install it again.

As it is now it:
  * looks to see if the file is versioned in git.svn. If it is not, it assumes it was copied from the server, and will be overwritten.
  * if the file was in version control, we assume the file was in the original repo, so we ignore the one from the server.

Another solution to this problem might be to do a `git clean` when updating, and this will remove all modifications from the user (including removing the original tmsbuild.yaml we put there). But it might be that we also remove modifications from the user he wanted to keep (not necessarily in tmsbuild.yaml but in any other file). So it looks a little bit safer to keep it without cleaning in update.