This test broke when we added vsoft.yaml, because it was testing installing vsoft.*.
Not the greatest design... but well, it should now be a little more robust, in that it finds out first how many vsoft.* products are. Still, it might break on the part that explicitly tests for the output:

Test-RepoFetchResultCounts -FetchResult $BuildResult -expectedLines @(
    "*- vsoft.commandline -> OK*",
    "*- vsoft.cancellationtoken -> OK*",
    "*- vsoft.yaml -> OK*")

We will have to add newer vsoft products here when we add them to the repo. Or just ignore this particular check in the test.
As we aren't adding vsoft products all the time, by the moment I believe the best is to keep the check, and update when a vsoft product is added. If it breaks all the time, we can remove it.