Fixes the -unregister flag in "tms build" and makes it public.
We already had a hidden "unregister" flag in tms build, but:

  * It was probably added before we had the ability to select particular products to build, so it would always unregister every product, even if you did `tms build myproduct -unregister`. See problem 1 in https://github.com/tmssoftware/tms-smartsetup/issues/214

  * It removed the hashes and everything except the product itself. That meant that if you unregistered a product and then did `tms build` it would rebuild the full thing again.

This fix attempts a more correct fix, instead of the shortcut of not loading the products. The advantages are:
   1 You can now unregister individual products
   2 When you unregister a product, and then call `tms build`, the product will be re-registered without any re-building.
   3 As an added bonus, smartsetup can now detect you changed the "skip registering" options. Before, if you had a product registered, edited the config to set skip-register=false and then call tms-build again, it wouldn't rebuild. Now it will.

The fix was a little more complex than what I wanted (I imagine that's why we didn't do it before), but I think it allows a new interesting workflow. Say, I could have 3 smartsetup installations with different products and versions:

smart\config1
smart\config2
smart\config3

I could have a batch file that calls tms build -unregister in config1 and config2 folders, and tms build in config3, then launches the ide. Now that register and unregister are instant operations, this becomes a possibility. You can activate and deactivate configs without having to rebuild the full thing (even if rebuilding it fast, just changing the lib paths and stuff is way faster).